+++ apps/ios/PhrenWidgets/PhrenWidgetsBundle.swift
@@ -6,5 +6,9 @@ struct PhrenWidgetsBundle: WidgetBundle {
     var body: some Widget {
         PhrenGlanceWidget()
         ApprovalActivityWidget()
+        SessionWorkingActivityWidget()
+        if #available(iOS 18.0, *) {
+            SessionAttentionControl()
+        }
     }
 }